+++ windows/flutter/generated_plugins.cmake
@@ -6,6 +6,7 @@ list(APPEND FLUTTER_PLUGIN_LIST
   audio_service_win
   just_audio_windows
   permission_handler_windows
+  share_plus
   url_launcher_windows
 )